+++ src/lib/project-tasks/bump-minor.ts
@@ -1,8 +1,8 @@
-import {bumpVersion} from "../utils/project";
-import {ProjectOptions} from "../config/config";
 import {Gulp} from "gulp";
+import {ProjectOptions} from "../config/config";
+import {bumpVersion} from "../utils/project";
 
-export const taskName = ":bump-minor";
+export const taskName: string = ":bump-minor";
 
 export interface Options {
   project: ProjectOptions;