+++ src/lib/project-tasks/index.ts
@@ -1,3 +1,4 @@
+import {Gulp} from "gulp";
 import * as bumpMajor from "./bump-major";
 import * as bumpMinor from "./bump-minor";
 import * as bumpPatch from "./bump-patch";
@@ -5,14 +6,14 @@ import * as install from "./install";
 import * as installNpm from "./install.npm";
 import * as installTypings from "./install.typings";
 import * as lint from "./lint";
-import {Gulp} from "gulp";
+import * as tslintJson from "./tslint-json";
 
 export interface Options extends bumpMajor.Options,
   bumpMinor.Options,
   bumpPatch.Options,
   install.Options,
-  lint.Options {
-}
+  lint.Options,
+  tslintJson.Options {}
 
 export function registerAll(gulp: Gulp, options: Options) {
   bumpMajor.registerTask(gulp, options);
@@ -20,6 +21,7 @@ export function registerAll(gulp: Gulp, options: Options) {
   bumpPatch.registerTask(gulp, options);
   install.registerTask(gulp, options);
   lint.registerTask(gulp, options);
+  tslintJson.registerTask(gulp, options);
 }
 
 export {
@@ -29,5 +31,6 @@ export {
   install,
   installNpm,
   installTypings,
-  lint
+  lint,
+  tslintJson
 };